requirements.txt
flask==3.0.0
flask-cors==6.0.1
pyjwt==2.8.0
werkzeug==3.0.1
sqlalchemy==2.0.23
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
jinja2>=3.0.0
requests==2.32.5
python-dotenv==1.0.0

# PDF Generation
reportlab>=4.0.0

# Admin panel
flask-admin==1.6.1
flask-wtf>=1.2.0
wtforms==3.1.1

# Document processing
pandas==2.1.4
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
markdown==3.5.1

# Optional: For better PDF processing
pymupdf==1.23.14

# OpenAI integration
openai>=1.3.0
requests>=2.25.1

# Azure Communication Services (Email)
azure-communication-email>=1.0.0,<2.0.0

# SAML SSO Authentication
python3-saml>=1.16.0
